This is a fairly big change and there may be bugs. Among many miscellaneous style changes and casual refactoring changes have been applied re: #43.
* Rather than asserting that the "homedir" setting has been set I added a
  default home and sub- directory to rnp_init(). Initialisation may still
  fail if that directory does not exist. The role of where the home
  directory needs to be required can be re-thought in subsequent
  changes.

* I removed the key loading responsibility from rnp_init().

* I created a new public library function, rnp_load_keys(). This should be
  called once when keys are due to be loaded. Unloading to protect against
  memory leaks may be introduced later.

The net effect of this is that where your code was previously:

```
apply_settings(rnp);
rnp_init(rnp);
```

it should now be:

```
rnp_init(rnp);
apply_settings(rnp);
rnp_load_keys(rnp);
```

In the interest of protecting your genuine .gnupg directory the commit
following this will change the default home subdirectory for gnupg keys
from .gnupg to .rnp. This change can be reverted after discussion.

This change is not yet ready for production and transitional changes will
need to be applied to dependent projects.