More noisy changes:
* Added a header for common macros to src/common/constants.h. Many of the
  bundled applications re-use common code, we may need to integrate a
  small private static library into their builds to make them easier
  to maintain.

* Updated hkpd, libpaa, libverify references to .gnupg subdirectory to use
  the common SUBDIRECTORY_* definitions.

* Updated hkpd, libpaa usage of rnp_init. rnp_load_keys was not yet added.

* Corrected some stylistic issues in each module and in some cases renamed
  netpgp to rnp.

Please note that I suspect that hkpd and libpaa utilities are not currently
being built. If I'm right there are probably bugs - you can shunt them
in my direction if you find any.

* Temporarily resolved the homedir directory/subdirectory issue. A more
  robust and permanent fix has been suggested in issue #44.

Currently rnpkeys still breaks due to the simplistic nature of the new
rnp_load_keys behaviour. This will be a simple fix that will arrive in the
next commit. If that passes testing I would suggest that this be merged
with name_refactor_wip.

Due to dependence on name_refactor_wip please note that this _should not be
merged to master_ until name_refactor_wip has been.