Add some continuous integration items. This uses Travis CI's beta container-based trusty build environment.
It will build with a few of clang's sanitize options and run the cmocka
tests. The tests will also fail if clang detects leaks, etc.

It will build and install botan at ~/builds/botan-install, and cmocka
at ~/builds/cmocka-install, and set appropriate variables to utilize them.
These builds are cached. So if, for example, we needed to build against
a newer version of Botan, we simply need to delete the cache on
travis-ci.org and restart the build.