Remove hash from reader.c.
This wasn't really used and was leaky.
The Hash: header in clearsigned messages is apparently used to indicate
the hash algo used in the signature (though it's redundant, as the signature
itself has ths information too).

We do still check to see if the hash algorithm is supported.

We should also record the header somewhere and verify later on
that it matches up with the signature's hash algo.
gpg checks for this and issues a "signature digest conflict" error.