Default to 2 CPU cores in CI scripts.
I previously overlooked the fact that /proc/cpuinfo is the host's
information, and is not adjusted for container limits or anything,
so we were sometimes trying to build things with 32
cores, when travis really tries to limit us to two.

Can be overridden (for example `env CORES=4 ci/install.sh`).