rnpkey: Generate ECDH key as subkey to ECDSA
ECDH key will always be signed with ECDSA key. ECDSA key
is a primary key and ECDH is created as a subkey of this key.

Additionally rnpkeys was generating EdDSA and SM2 keys with
same subkeys, which obviously is wrong. I'm not sure if any subkey
should be created when user requests EdDSA or SM2, but definitelly
EdDSA must be used for encryption.
With this patch rnpkeys will always create default (RSA-2048) subkey
for EdDSA.