repgp: Check usage flag before encrypting
Code was trying to encrypt always with primary key. Primary keys
are mostly used for signing.
With this patch code will check if key used for encryption
has PGP_KF_ENCRYPT flag, if not it will try to look in subkeys
for encryption key. It will fail if proper key is not found