Improve signature tampering in testSignature test.
This test was failing when tampering with an armored signature when the
rnp Version: header was a little bit longer, because it would end up
just modifying the header (not the signature data), so the verification
would pass even when the test believed it should not. Tampering with the
middle byte of the signature data seems like a slightly better approach.