ci: Use shallow clones.
We don't really need the full history on any of these. Right now a
cache rebuild for GPG_VERSION=beta will git clone about ~800MiB. This
cuts that in half so it's probably worth doing.

Only thing to be aware of is that --branch will try a remote branch name
first, and a tag second. But I can't see any issue with that for us (and
generally people avoid naming tags and branches identically for reasons
like this).