ECDSA: pass expected hash length via "Raw(hashname)" param
This causes the matching hash to be used for the deterministic
nonce generation. It also checks that the input is of the expected
length.

This removes the code to truncate the ECDSA input as since the merge
of https://github.com/randombit/botan/pull/1502/ Botan handles it
already. This patch was included in Botan 2.5 which is already the
minimum required version.

GH #367