Make invoking GnuPG more robust.
GnuPG will try to read its state even for operations that do not
require it, e.g. listing packets.  If the state directory does not
exist, GnuPG will try to create it.  If this fails, GnuPG errors out.

In some build environments, $HOME may not exist or may not be
writable (e.g. in Guix and Nix, $HOME does not exist).  This leads to
a spurious test failures when GnuPG is invoked to inspect packets.

Fix this by using the current directory as GnuPG state directory.