Avoid embedding major version number in library name
The indication of API and ABI stability for a C library is the SONAME,
which already includes the necessary versioning information.

We don't want users of librnp to try to build against two versions of
the library at once, so there isn't a strong reason to have two copies
of devel info (headers/pkg-config/etc) installed at once.

Two different binary shared objects with SONAMEs should already be
able to be installed concurrently, so this doesn't affect the
situation with multiple distinct tools on a system that are built
against different API or ABIs.

Closes: #1406