ci: Make Cirrus CI skip CI also when 'skip ci' occurs elsewhere in the body of the commit message
The doc at https://cirrus-ci.org/guide/writing-tasks/#conditional-task-execution
says that it would skip CI if 'skip ci' is on the first line of the
commit message, but this behaviour would not match configs found in
GitHub workflows.

This commit tries to match their behaviours.

It would also not run Cirrus CI if the only changes are in .github
workflows.

[skip ci]