Enable loosening the s2k_iteration_tuning test
Some platforms or some specific machiens (e.g. the s390x and sparc64
debian build daemons) don't appear to consume a linearly-proportional
amount of wall clock time as the S2K iteration count increases.

This patch changes nothing substantive about the s2k_iteration_tuning
test, but enables specific machines or architectures to override the
expectation with a simple #define (e.g. by adding
-DS2K_MINIMUM_TUNING_RATIO=1.5 to CXXFLAGS).

I'm generally of the opinion that this kind of tuning is misleading or
even counterproductive when your threat model is modern hardware (the
iterated S2K count octet should probably just always be maxed out to
0xff, and even better would be to stop using this S2K mechanism and
adopt the upcoming argon feature).  So the right thing overall might
be to stop doing this tuning entirely.

But this targeted change would be sufficient for debian's
purposes. With this patch, we can easily tell specific build daemons
that they don't need to worry about a linear ratio between 10
iterations and 100 iterations of SHA512 and get on with the rest of
the test suite.

Closes: #1439