Add a codespell configuration file
This makes it relatively easy to avoid introducing new common,
detectable spelling errors by running:

    codespell *

Note: it still highlights a misspelling of MERCHANTABILITY in
LICENSE-OCB.md, but since that appears to be part of a signed legal
agreement i didn't want to touch it.

Probably easiest to add it to the comma-separated "skip" list if you
want to just have it leave that file alone going forward.