CLI tests: drop a strange behaviour of file_text() utility function
file_text() is a function in src/tests/cli_common.py which reads the
whole file (in binary mode, so it doesn't transform newlines in
OS-dependent way) and returns it as a Python string, making a two-liner
task a single function call.

But it also does `.replace('\r\r', '\r')` on the resulting string, which
is questionable for a function which has a very general-purpose name,
doesn't have this behaviour documented, and is used in 36 places across
the test suite. It's unclear what was the original intent, and whether
this was a series of typos or not (`\r\n -> \n`?).

Dropping this transformation altogether doesn't affect any of existing
tests.