tests: relax gpg handling of partial-length public keys
gpg is expected to reject public key packets that are partial length.
But at least some implementations skip over malformed pubkeys in
the keyring without causing an error.

Confirm that the output of `gpg --list-keys` on a pubring.gpg that
contains only a partial-length public key is empty, rather than
expecting it to return non-zero.

See also:

   https://gitlab.com/sequoia-pgp/sequoia-chameleon-gnupg/-/issues/148